+++ CHANGELOG.md
@@ -6,8 +6,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## \[2.1.0] - Unreleased
 
+* \[Feature] - Added `HelpArg` and `VerboseArg` mixin
 * \[Feature] - Added `IShell.throwsOnError()`
-* \[Test] - `PocessRunShell` Tests
+* \[Test] - `ProcessRunShell` Tests
+* \[Deprecated] - `HelpOption` and `VerboseOption`
 
 ## \[2.0.2] - 2022-02-10